insurance-data-generator/requirements.txt
# Azure SDK
azure-eventhub==5.11.5
azure-eventhub-checkpointstoreblob==1.1.4
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Data Generation
faker==22.2.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4

# UK-specific
ukpostcodeparser==1.1.2
geopy==2.4.1

# Streaming & Async
asyncio-throttle==1.0.2
aiofiles==23.2.1

# API & Dashboard
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0
folium==0.15.1
plotly==5.18.0
streamlit-folium==0.17.4

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
colorlog==6.8.0
tenacity==8.2.3